requirements.txt
streamlit
torch
torchvision
matplotlib
Pillow
numpy
gdown
ultralytics